removes old states
The states were used when there was more than one salt process run.  Now
that there is only one run, and the app gets picked up for
statefullness, we can use the server’s states again.  In this case
mysql-start replaces mysqld.  I could have checked for the services, but
i would need to make sure that there was not an ID collision