one last clear of the cache
on the first load, what happens is that due to the past installations of
the extensions, there is a cache that is already build up.  This is
normal, but when we do the addition of the sample sites, they are not
caches and need to be picked up.  This is done by clearing the cache so
Magento reacts correctly.